refact(): delete some useless files and fix(): fixing home page
