+++ frontend/src/app/app.module.ts
@@ -6,16 +6,17 @@ import { FormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
-import { HomeModule } from './home/home.module';
+
 import { HistoricListComponent } from './historic-list/historic-list.component';
 import { FollowersComponent } from './followers/followers.component';
 import { ProfileModule } from './profile/profile.module';
-import { NotFoundComponent } from './not-found/not-found.component';
 import { FeedComponent } from './feed/feed.component';
-import { FindComponent } from './find/find.component';
+
+import { HomeModule } from './home/home.module';
+import { NotFoundComponent } from './not-found/not-found.component';
 
 @NgModule({
-    declarations: [AppComponent, HistoricListComponent, NotFoundComponent, FindComponent],
+    declarations: [AppComponent, HistoricListComponent, NotFoundComponent],
     imports: [
         FormsModule,
         BrowserModule,
