+++ frontend/src/app/find/find.component.html
@@ -1,5 +1,6 @@
 <div class="search-container">
-    <input type="text" [(ngModel)]="query" (input)="onSearch()" placeholder="Search users..." />
+    <input type="text" [(ngModel)]="query" (input)="onSearch()" placeholder="Pesquise usuários..."
+           class="search-input" />
     <div *ngIf="users.length" class="search-results">
         <div *ngFor="let user of users" (click)="onUserClick(user.id)" class="user">
             <img [src]="_getProfileImage(user.id)" alt="Foto do perfil" class="profile-picture" />
