+++ frontend/src/app/find/find.component.scss
@@ -1,6 +1,6 @@
 .search-container {
     width: 300px;
-    margin: 0 auto;
+    // margin: 0 auto;
     padding: 20px;
     background-color: #f0f0f0;
     border-radius: 5px;
@@ -25,6 +25,7 @@
     display: flex;
     align-items: center;
     margin-bottom: 10px;
+    cursor: pointer;
 }
 
 .profile-picture {
