--- frontend/src/app/home/components/create-item/create-item.component.scss
@@ -1,18 +1,0 @@
-.card {
-    background-color: var(--color-cards);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    text-align: center;
-    padding: 3rem;
-    gap: 0.6rem;
-
-    .title {
-        font-size: 2.5rem;
-        font-weight: bold;
-    }
-
-    .input {
-        margin-top: 2rem;
-    }
-}
