--- frontend/src/app/home/components/create-item/create-item.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CreateItemComponent } from './create-item.component';
-
-// TODO: fix this test
-
-describe('CreateItemComponent', () => {
-    let component: CreateItemComponent;
-    let fixture: ComponentFixture<CreateItemComponent>;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({
-            declarations: [CreateItemComponent],
-        });
-        fixture = TestBed.createComponent(CreateItemComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
