--- frontend/src/app/home/components/list-items/list-items.component.html
@@ -1,8 +1,0 @@
-<mat-list class="list" role="list">
-    <mat-list-item
-        class="item"
-        role="listitem"
-        *ngFor="let item of items; let i = index">
-        <span [dataCy]="'item-' + item.id">{{ i + 1 }} - {{ item.name }}</span>
-    </mat-list-item>
-</mat-list>
