--- frontend/src/app/home/components/list-items/list-items.component.scss
@@ -1,15 +1,0 @@
-.list {
-    background-color: var(--color-background);
-    display: flex;
-    flex-direction: column;
-
-    .item {
-        display: flex;
-        text-align: left;
-        padding: 0;
-
-        span {
-            font-size: 1.5rem;
-        }
-    }
-}
