--- frontend/src/app/home/components/list-items/list-items.component.ts
@@ -1,11 +1,0 @@
-import { Component, EventEmitter, Input } from '@angular/core';
-import { Item } from '../../types/item';
-
-@Component({
-    selector: 'app-list-items',
-    templateUrl: './list-items.component.html',
-    styleUrls: ['./list-items.component.scss'],
-})
-export class ListItemsComponent {
-    @Input() items: Item[] | null = [];
-}
