+++ frontend/src/app/home/home-routing.module.ts
@@ -1,17 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './pages/home/home.component';
-import { ItemsComponent } from './pages/items/items.component';
+import { HomeComponent } from './home.component';
 
 const routes: Routes = [
     {
         path: '',
         component: HomeComponent,
     },
-    {
-        path: 'items',
-        component: ItemsComponent,
-    },
+    // {
+    //     path: 'items',
+    //     component: ItemsComponent,
+    // },
 ];
 
 @NgModule({
