+++ frontend/src/app/home/home.component.html
@@ -1,0 +1,17 @@
+<div class="home-page"> 
+    <!-- <img src="../../assets/background/13.jpg" alt="Logo da Epic Games" class="logo-epic-games"> -->
+    <div class="header">
+        <div class="logo">Jogos Épicos</div>
+    </div>
+    
+    <div class="centered-container">
+        <app-find></app-find>
+    
+    </div>
+    <!-- Colocar rota para cadastro -->
+    <a class="link" routerLink="/">Cadastre-se</a>
+</div>
+
+
+
+
