+++ frontend/src/app/home/home.component.scss
@@ -1,0 +1,39 @@
+.home-page {
+    background-color: rgb(22, 8, 80);
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    height: 100vh; /* Define a altura para ocupar toda a altura da tela */
+    background-image: url("../../assets/background/13.jpg");
+}
+
+.header {
+    position: fixed;
+    // padding: 120px 580px; /* Espaçamento interno */
+    color: white; /* Cor do texto */
+    font-size: 100px; /* Tamanho da fonte */
+    font-weight: bold; /* Peso da fonte */
+    margin-top: 5%;
+}
+
+.centered-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh; /* Define a altura para ocupar toda a altura da tela */
+    top: 0;
+}
+
+.link {
+    text-decoration: none;
+    color: white; 
+    margin-bottom: 5%;
+    font-size: 200%;
+    cursor: pointer;
+}
+
+// Hover link
+.link:hover {
+    color: rgb(126, 126, 126);
+}
