+++ frontend/src/app/home/home.component.ts
@@ -1,0 +1,13 @@
+import { Component, HostListener } from '@angular/core';
+import { Router } from '@angular/router';
+
+@Component({
+    selector: 'app-home',
+    templateUrl: './home.component.html',
+    styleUrls: ['./home.component.scss'],
+})
+export class HomeComponent {
+
+    
+
+}
