--- frontend/src/app/home/pages/home/home.component.html
@@ -1,6 +1,0 @@
-<div class="container">
-    <app-create-item
-        (goToListEvent)="goToList()"
-        (createItemEvent)="createItem($event)">
-    </app-create-item>
-</div>
