--- frontend/src/app/home/pages/home/home.component.scss
@@ -1,7 +1,0 @@
-.container {
-    background-color: var(--color-main);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-}
