--- frontend/src/app/home/pages/items/items.component.scss
@@ -1,18 +1,0 @@
-.container {
-    background-color: var(--color-main);
-    display: flex;
-    align-items: center;
-    text-align: center;
-    justify-content: center;
-    height: 100%;
-
-    .card {
-        padding: 3rem;
-
-        .title {
-            font-size: 2.5rem;
-            font-weight: bold;
-            margin-bottom: 1rem;
-        }
-    }
-}
