--- frontend/src/app/home/pages/items/items.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ItemsComponent } from './items.component';
-
-// TODO: fix this test
-
-describe('ItemsComponent', () => {
-    let component: ItemsComponent;
-    let fixture: ComponentFixture<ItemsComponent>;
-
-    beforeEach(() => {
-        TestBed.configureTestingModule({
-            declarations: [ItemsComponent],
-        });
-        fixture = TestBed.createComponent(ItemsComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
