--- frontend/src/app/home/types/item.ts
@@ -1,4 +1,0 @@
-export type Item = {
-    id: string;
-    name: string;
-};
