+++ frontend/src/assets/getImages.service.ts
@@ -1,10 +1,10 @@
 function getProfileImage(prefix_path: string, id: number): string {
-    id = id % 99;
+    id = id % 98;
     return `/assets/profile/${id}.jpg`;
 }
 
 function getBackgroundImage(prefix_path: string,  id: number): string {
-    id = id % 23;
+    id = id % 19;
     return `/assets/background/${id}.jpg`;
 }
